implementação do banco abstrato e .env para utilização de varios tipos de banco no projeto

+++ src/model/DAO/InterfaceDAO.java
@@ -6,7 +6,7 @@ public interface InterfaceDAO<T> {
     
     abstract void create(T objeto);
     abstract List<T> read();
-    abstract T read(int codigo);
+    abstract T read(Long codigo);
     abstract void update(T objeto);
     abstract void delete(T objeto);
 }
